Read config files during server startup, instead of once per request
