+++ api/data_explorer/controllers/dataset_controller.py
@@ -1,10 +1,9 @@
 import connexion
 import six
 
+from flask import current_app
+
 from data_explorer.models.dataset_response import DatasetResponse  # noqa: E501
-from data_explorer import util
-from ..dataset_faceted_search import get_dataset_name
-from ..dataset_faceted_search import get_table_names
 
 
 def dataset_get():  # noqa: E501
@@ -15,4 +14,5 @@ def dataset_get():  # noqa: E501
     :rtype: DatasetResponse
     """
     return DatasetResponse(
-        name=get_dataset_name(), table_names=get_table_names())
+        name=current_app.config['DATASET_NAME'],
+        table_names=current_app.config['TABLE_NAMES'])
