+++ api/data_explorer/test/test_dataset_controller.py
@@ -2,6 +2,14 @@ from flask import json
 
 from data_explorer.test.base_test_case import BaseTestCase
 
+dataset_name = 'Test data'
+table_names = [
+    "test_project1.test_dataset1.test_table1",
+    "test_project1.test_dataset1.test_table2",
+    "test_project2.test_dataset2.test_table3",
+    "test_project2.test_dataset3.test_table4"
+]
+
 
 class TestDatasetController(BaseTestCase):
     """DatasetController integration test stubs"""
@@ -9,7 +17,8 @@ class TestDatasetController(BaseTestCase):
     def create_app(self):
         app = super(TestDatasetController, self).create_app()
         app.config.update({
-            'DATASET_CONFIG_DIR': 'dataset_config/test_dataset',
+            'DATASET_NAME': dataset_name,
+            'TABLE_NAMES': table_names
         })
         return app
 
@@ -18,14 +27,7 @@ class TestDatasetController(BaseTestCase):
         response = self.client.get('/dataset')
         self.assert200(response)
         self.assertEquals(
-            dict(
-                name='Test data',
-                tableNames=[
-                    "test_project1.test_dataset1.test_table1",
-                    "test_project1.test_dataset1.test_table2",
-                    "test_project2.test_dataset2.test_table3",
-                    "test_project2.test_dataset3.test_table4"
-                ]), response.json)
+            dict(name=dataset_name, tableNames=table_names), response.json)
 
 
 if __name__ == '__main__':
